We are now failing because the `init` call that triggers flag creation & lookup happens before we register the custom type handlers.
I think the most straight-forward way to deal with this
(and not have to think about initialization order)
is to change the `goat.Register` call.
Instead of registering a run config,
we'll register a function that returns a run config.
Then we'll only run it when we need to execute it,
which will be _after_ we register our custom types.